Resolves #697: OnlineIndexer should make the transaction priority configurable
This adds a `setWeakReadSemantics` and a `setPriority` method to the `OnlineIndexer.Builder` method. Adding a way to set the weak read semantics was simple, as it was already a method on runners. To thread the priority through, I added a priority field to transactions and to runners, and then the online indexer exposes that. The `OnlineIndexer.Builder` method then sets the priority to `BATCH` (by default).

I was tempted to make the default weak read semantics one that set causal read risky to true. This would make sense for the online indexer, as it can actually tolerate stale reads (as everything is committed anyway), but it also benefits from seeing its own writes, which causal read risky actually guarantees. However, that is technically a behavior change, though it should be easy to add later if desired.

This is on top of #698 because (a) they conflict otherwise and (b) there are a few methods in #698 that this can use. It could use the underlying FDB methods, but I would prefer it use the new one instead. The extra instrumentation is also useful for determining if setting weak read semantics on the index build actually did anything as I could look at the GRV instrumentation.