Fixes #705: Truncate transaction IDs to 100 characters
This adds a "sanitization" step to `FDBRecordContext` before it will set the transaction identifier. If the identifier exceeds 100 characters, it will truncate or drop the string. This operates on the assumption that the ID here is less important to most users than actually executing the operation. This also allows the user to set whether to log the transaction separately from the call in the `FDBRecordContext` constructor. One could already do that by calling `ensureActive`, but this provides a first-class function for that.