Resolves #707: Add an existence check option that only warns if there are no records
This new existence check will only warn if the first key in the record store (at `checkVersion` time) is not a record. As this is a sign of corruption, it is not the default and it still logs at warn level. However, as this is recoverable as the record store can just operate as if the store is new, it is safer to use this option rather than `NONE`. It will organically repair stores as it finds them, assuming that the operation commits.

I went with a whitelist approach for the keyspaces rather than a black list. (For example, an alternative approach could be to check if the first key is in the "records" or "indexes" subspace.) I think the whitelist of allowable keyspaces for the first key is safer, though it might mean that we need to update that place in the codebase as we add more keyspaces.