instrument grv at store open time
It is possible that this will miss a few cases that do not call getReadVersionAsync, but I think this covers the main cases. In the case where an existing user has code that calls getReadVersion on the transaction, this will add a small amount of overhead to go back to the network thread and fill in the value from FDB. In the case where a user calls getReadVersion on a record context, then this adds an even smaller amount of overhead as the function will return a completed future anyway.

Also, some typos.