Resolves 731: Reload online indexer's rebuild configuration parameters
The configuration paremeters of online indexer used to be passed only in
the beggining of the index build. However, these parameters may change during
long builds and never be picked up by the indexer. This change provides
a way for the indexer to load the config parameters in the beginning of
every transaction.