Add tests for #744.
This adds test to demonstrate a deficiency in the planner pointed out in #744. The failing tests are disabled here so that they do not break the build but so that (a) there is a record in the code of the kind of cases that are broken here and (b) once the problem is fixed that PR can just enable these tests. This also makes the RecordTypeKeyTest extend from `RecordQueryTestBase` so that the `@DualPlannerTest` annotations can be applied, though it does not actually make any tests use the new planner (as they appear to not work yet with the way the planner is on master).