Resolves 731: Reload online indexer's rebuild configuration parameters (#732)
* Resolves 731: Reload online indexer's rebuild configuration parameters

The configuration paremeters of online indexer used to be passed only in
the beggining of the index build. However, these parameters may change during
long builds and never be picked up by the indexer. This change provides
a way for the indexer to load the config parameters in the beginning of
every transaction.

* Update docs/ReleaseNotes.md

Co-Authored-By: Alec Grieser <alloc@apple.com>

* Update fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexer.java

Co-Authored-By: Alec Grieser <alloc@apple.com>

* Addressed Alec's feedback

* Applied Tao's suggestion and fixed a spotBug issue

* Assert increaseLimitAfter and progressLogIntervalMillis remain unchanged

* Addressed Alec's comments