Fixes #765: Inequality comparisons are not coalesced on single field indexes
This makes an update to the old planner to support queries with two inequality comparisons on a single range. This extends the `AndWithThenPlanner` to also work on indexes that are actually on a single field, and then it uses that logic in `planAnd`. The alternative would be to duplicate the logic there, which seems wrong. In theory, more code could be condensed to the `AndWithThenPlanner` (which probably needs a new name).