Fixes #763: Evaluating a query component for equality comparison with a enum constant fails
The approach here was to compare enum fields based on their field number. This should give the same semantics for enum equality and inequality comparisons when no index is defined as is done when there is an index. Their is some amount of type safety lost in the comparison, which is unfortunate.