Resolves #702: Increase required FDB version to 6.2
This increases the required FDB client and server version to 6.2, and it updates our build infrastructure to use that newer version. The main feature that this will enable is it will allow us to use the new `getApproximateSize` feature in 6.2, but the work to use this in, say, index builds has not been done.