Resolves #839: Add aggregate counters for the number of reads and writes
In addition to the current metrics we have on the number of bytes read, written, and deleted, this also adds aggregates for the number of keys read, written, and deleted (etc.). This can be useful as there are some pathologies that result from lots of small reads, so having something that tracks the number of distinct reads can be useful.