Resolves #841: Add a custom index queryability filter for planning.
Add a new field to RecordQuery that allows a user to specify a custom
filtering function that determines whether an index should be considered
by the planner. This replaces the previous hard-coded check for the value
of IndexOptions.ALLOWED_FOR_QUERY_OPTION, which remains the default.

This change does not break backwards compatibility.