Resolves #864: KeySpacePath resolution should re-use the initiating transaction's read version
This adds overloads to `LocatableResolver::resolve` that allow a transaction to be passed, and then the resolver uses that transaction's read version when performing reads from the directory layer. In the "usual case" (which should be when a new entry does not need to be created), this transaction won't be committed anyway, so the only loss in efficiency over using the transaction itself is some extra memory (and no extra network requests).

It would probably still be a good idea to do #853 (if we can figure out what we want to do about caching), but this is easier and gets us 99% of the benefit. That being said, there is more cognitive load needed to understand this API, which is perhaps a smell.