Fixes #876: TextIndex does not instrument BunchedMap reads and writes
The problem here was that the reading and writing of keys from FDB happened far enough away from the store timer logic, etc., that the previous logic was not even trying to look within the BunchedMap. However, quite a few reads and writes could happen as a result of text indexing, so getting the right numbers is fairly important. Note that once #866 is in, this may matter less (as the total reads and writes for the transaction will be more accurate), but doing it this way allows it as well allows it to (1) include the sizes of deleted values and (2) correctly correlate those operations to index updates. For that reason, I think this has value even after that other issue is addressed.