Resolves #881: Index state reads should happen with a WANT_ALL streaming mode
This configures the read of the index states as part of check version to read all the states using the `WANT_ALL` streaming mode. This should decrease the number of round trips to read all states. Note that we were already waiting for all results to come back anyway, so there is not the downside one sometimes get from using `WANT_ALL` where the time to first byte is lower.