Revert #885: Re-enables fdb-record-layer-spatial builds
This re-enables building and publishing the `fdb-record-layer-spatial` library. However, I think we should leave #883 open for now, as I think there is a desire to make the builds more resilient to these kinds of failures in the future, but there is no reason not to let them continue for now.