Fixes #888: IN to OR transformation does not plan with other predicates.
PR #861 added a new transformation to the RecordQueryPlanner that
attempts to transform an IN predicate into an equivalent OR of equality
predicates when the planner determines that it cannot implement the IN
predicate as an IN-join because of an incompatible sort order.

However, the equivalent OR predicate is too complex for the simple
normalizer normalizeAndOr() if there are other predicates on the
original filter. In this case, the predicates are not normalized and so
the union planner is unable to produce the desired union plan.

This change adds a more sophisticated normalizer to handle this case.