Test: Don't use openStore helpers just to open something else
There's a bunch of test code that calls openStore, and then converts
it to a builder, to change something and then re-open. Inline those,
in part to deal with the upcoming fix to #489, and in part because
it makes the code cleaner.
Also, one of the tests did not have a try around the contexts.