Fixes #482: Test failure in RankIndexTest::repeatedRankQuery
This looks like it was a real concurrency bug (one that can lead to the data being corrupt). Essentially, the ranked set does not allow concurrent updates to the same subspace in the same transaction. This is a known issue, and there are real ways to fix it (like implementing some kind of locking abstraction), but it is hard to get right in a way that also does not result in blocking in async threads. The patch here fixes the text index maintainer so that it issues its updates to the same subspace serially, though it can still update unrelated subspaces in parallel.