Resolves #925: Add assertions into record layer
This change allows `FDBDatabaseFactory` to be configured with a supplier which
can decide whether or not assertions are to be enabled within the record layer
API.  This supplier is polled each time an `FDBRecordContext` is created and,
thus, all operations from the context will have assertions checked (or not)
as it is used.  The initial set of assertions are around validating key and
value byte sizes.