Fixes #901: False positive of uniqueness violations prevents building indexes (#902)
* Fixes #901: False positive of uniqueness violations prevents building indexes

* Move uniqueness APIs down to StandardIndexMaintainer & Fixed a bug in removeUniquenessViolationsAsync & Improved tests

The main problem fixed by this PR is that: `FDBRecordStore.resolveUniquenessViolation` used to be relied on to delete the last uniqueness violation entry when delete the second last one. But this is not the only way the uniqueness violation is resolved. It doesn't cover the case that the violation is resolved when record is deleted organically. To resolve this problem, now it will delete the last entry as part of deleting records (See `removeUniquenessViolationsAsync`).

This change also tries to consolidate uniqueness logic at `StandardIndexMaintainer` level:
- `FDBRecordStore .addUniquenessCheck`  -> `StandardIndexMaintainer.checkUniqueness`
- `IndexMaintainer.updateUniquenessViolations` -> `StandardIndexMaintainer .addUniquenessViolations` and `StandardIndexMaintainer.removeUniquenessViolationsAsync`