Fixes #952: Local version cache is keyed only by Tuple (i.e., primary key) not subspace
This changes the local version cache to be keyed by `byte[]` rather than by `Tuple`. Then the exact key (in the database) is used to represent the cache entry rather than the record's primary key. This allows multiple stores to be opened in the same transaction and for them to interact with versioned records with the same primary key in each store without having cache entries from one store pollute the cache entries of the other.

This meant changing the API. The previous API was just wrong, in that there was absolutely no way to protect oneself from the bug identified in #952, so it has just been removed (despite the fact that this violates are claims of API stability). These APIs probably should have been marked INTERNAL anyway, but I believe they pre-date the API stability annotations, which is why they weren't.