Resolves #958: VersionstampSaveBehavior could have option to match version argument precisely
This adds a new `VersionstampSaveBehavior` setting, `IF_PRESENT`, which only saves a version with a record if the provided version is not `null`.
This is useful for times when you want to, say, copy the data from one store to another precisely, so the default behavior, where creating a
a version if not set explicitly, isn't desired.

In doing this, I also filed #964 because the test I wrote seemed to expose this. It's possible this behavior is correct "enough", but I'm
not super happy with #964 as it is right now.

This resolves #958.