Fixes #974: OrElseCursor doesn't properly support continuations.
Use a new structured continuation object and protobuf serialization to
record the complete state of the OrElseCursor, including which branch
was chosen. Rewrite OrElseCursor's logic in a simpler state-machine
style, which makes it easy to resume the cursor before it has made a
decision about which branch to take.

To avoid problems like we saw with the FlatMapPipelinedCursor's
continuations, deprecate the RecordCursor.orElse() fluent method and
replace it with a static method that interprets the continuation.

Because this changes the continuation format, this is a breaking change.