Resolves #991: Perform record fetches after union/intersection.
Implements a delicate-but-working pushdown of union, intersection, and
primary key distinct operations below the record fetch (from an index
scan) within the RecordQueryPlanner.

This is a fairly delicate operation that isn't easy to express properly
within the RecordQueryPlanner. In a number of places, the transformation
as expressed is incorrect, but happens to be correct in all the
circumstances where we happen to use it.

Because the execution framework cannot currently manipulate streams of
objects other than records, index entries are represented as partial
records, as in covering index scans.