Resolves #1016: Covering filters on legs of unions/intersections.
In #1001, we added an optimization that allows filters to be executed
on partial records produced by covering index scans. In #992, we also
added the ability to perform union/intersection/primary key distinct
operations before fetching the base records by primary key.

This refactors the work from #1001 into a substitution visitor that can
work in concert with the union/intersection/distinct visitors. It also
corrects some (untriggered) bugs in the existing visitors.

This also fixes #1017; full text fields now require looking at the
primary records and are not "satisfied" by the tokens in the full text
index entries.