Fixes #1032: In-as-or optimization can infinitely recurse if normalization as "OR" preserves
This fixes the infinite recursion by noticing when the in-as-or optimization cannot be applied (as the filter is unchanged by extracting the IN). This relies on the fact that that transformation is idempotent, so even if the query was not satisfiable by any index even after the sort, going through this transformation should stop it from infinitely recursing.