Fixes #1046: Setting a time limit can lead to "split record without start" error when reading data
The issue here is that in the code as it was originally, if the meta-data does not allow split records, then the time limit was enforced on the key value cursor. (If you had split records, then the time limit was removed on the very bottom key value cursor, and then only enforced in the wrapping cursor (the "key value unsplitter").) This meant that if you had versionstamps enabled, it was possible for the limiter to stop midway through, between a record and its version (if the time limit was hit at just the wrong time). Now, this removes the time limit from the key value cursor (always) and instead enforces that limit in the wrapping cursor so that it will always stop after reading a full record. This was already being done for the other kinds of limits, but not the time limit.