Resolves #1102: Regenerate .idea/compiler.xml
This regenerates the `.idea` configuration for the compiler. This was done by just importing the probject into the IDE and then letting it edit the file. This basically just sets the dependencies, etc., and a few language target things. We may have to regenerate this file again, but this should stop this change from being accidentally included in future PRs.