Resolves #949: Useful system keys should be exposed through a more useful API (#979)
* Resolves #949: Useful system keys should be exposed through a more useful API

This creates a new set of constants for the system keys that we care about, and then it changes the instances
that were already referencing system keys, and then I added a new class, `FDBSystemOperations`, that allows
a user to use these features.

I'm not...super happy with this API, though. Unlike many operations, it wasn't sufficient to provide a context
for some of these operations, as you might need to set some extra options. But then to provide the user
a way to provide their own retry policies (and things like a timer, etc.), it felt like the least bad way
to pass that information, though we don't use a runner like this in other contexts.

This resolves #949.

* Apply suggestions from code review

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>

* Remove public methods that take an `FDBRecordContext` from `FDBSystemOperations`. Clarify a few comments, etc.

* Update fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/FDBSystemOperations.java

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>