Resolves #1213: New indexes should generally begin life as DISABLED instead of WRITE_ONLY if not READABLE
This updates the default value in the check version logic and also in the `UserVersionChecker`'s default implementation of `needRebuildIndex`. It leaves the old method around so that users can opt in to the old behavior by providing a `UserVersionChecker` with that old method around. Javadoc has been added in a few places to tell the user what the point of that method is and what the return value means, and also points them to those existing static utility methods that they can use within their implementations. The `OnlineIndexer` already had a default `IndexStatePrecondition` that was compatible with this change, so I mentioned in the release notes how one can change it if one is using the only incompatible `IndexStatePrecondition`. The other changes are to tests, usually to ones that were explicitly testing that an index began as `WRITE_ONLY`, but they usually really only cared that the index wasn't `READABLE`, as that would imply a bug where indexes were being made `READABLE` before they were built. But the tests didn't _actually_ care about the difference between an index being `DISABLED` or `WRITE_ONLY`.