Fixes #1258: FDBReverseDirectoryCache::putIfExists can throw if using a shared read version
This adds a test that simulates a problem that was seen with the reverse directory cache. Essentially, if two transactions picked the same value for a name (with the same scoped), then we could fail that operation with a non-retriable error due to some checks that looked at an in-memory cache. The proposed fix here is to make the exception retriable. I added a test that simulates creating a bunch of entries all with the same transaction, and so probabilistically, some of them should allocate to the same value.