Resolves #1229: Expose range size estimate on Record Stores
This exposes the range size methods through the `FDBRecordStore` API. One is added for the whole store and another for all records. In theory, we could also add one for something like checking the size of an individual index or for ranges of keys. The latter might be useful when we go to #7, as we could use it for certain record stores for estimating the size of a single record type (if the store's primary keys are all grouped by the record type).