Fixes #1281: AutoService missing from Record Layer runtime dependencies (#1282)
Fixes #1281: AutoService missing from Record Layer runtime dependencies

This reverts our `AutoService` dependency to 1.0-rc4, which was the previous version. At that version, the `AutoService` annotation is retained at `SOURCE` only, which means that it is not included in the generated `.class` files, so our downstream dependencies don't need to worry about what version of `AutoService` we're using or even that we're using `AutoService` at all (and are free to use their own version or an alternative mechanism for adding to the service registry entirely). However, with 1.0-rc7, the annotation became a `CLASS` retained object, which means that it would need to start being something our downstream dependencies would need to worry about. To avoid putting that onto our dependencies, this just reverts the upgrade. I'm not excited about us being stuck on this older version, but I think it's more expedient, at least until we can get Java 11 modules all worked out.

This fixes #1281.