Fixes #1294: Scanning a store with a limit can result in "attempted to return a result with NoNextReason of SOURCE_EXHAUSTED but a non-end continuation"
This adds a guard to checkout for the an exhausted inner cursor, and if that happens, the scanning logic will now return an exhausted record cursor instead of trying to create a result with `SOURCE_EXHAUSTED` as the no-next-reason and a non-end continuation. Note that this bug requires that the final record in a store be split record with multiple parts and the out-of-band limit has to line up with hitting that record, so in practice, this should be a fairly rare occurrence, even though I could write a test that hit this 100% of the time.