Resolves #1239: Use either record count or estimated size when determining IndexStatus in checkVersion
This allows library consumers to create a `UserVersionChecker` that inspects the size of the record store rather than the number of records, which may be useful for stores that don't have a record size index or for whom determing the record count is more work than should be done in a single transaction (see #7). It also makes getting the count or size lazy, and so if the `UserVersionChecker` can determine an index state without needing to read from the database (such as if the index is on new types), it can avoid having to do I/O entirely.