Resolves #1304: Expose transaction option setServerRequestTracing in context config
This adds a configuration option to transactions that allows server request tracing to be enabled for that transaction. The ergonomics at the moment are a little bit weird, as if you set the option and provide a transaction identifier, then there's a single trace log message saying which random trace ID correlates with that transaction ID. However, individual sub-components of the transaction will have their own IDs (partially necessary to support many-to-one relationships between parents and children, which is required to accurately represent the effects of batching within FDB internals), so that only gets you so far. But this is what's required to make what is currently exposed work.