Resolves #1315: Allow indexes to be marked as replacing existing indexes
This adds a new index option prefix, `replacedBy`. This then allows the user to use the prefix to specify one or more indexes, and when all of those indexes have been built, the index will be disabled and its data deleted. This allows the user to effectively drop an indexe, though only after when the index has been fully replaced by new indexes. This both allows older unused data to be protected while also protecting the user against removing the index before all of the queries that use it have been redirected to other indexes.