Respond to Mike's comments (sort of)
After doing some thinking, I decided my former approach was misguided. The underlying goal
was to be able to capture I/O distribution information, however the approach that
I had would ultimately not allow this.

This changes tacts and introduces the ability to install a listener in the
database factory that will be notified each time a transaction is committed or
closed and will be provided metrics about the I/O done by the transaction.
This can be trivially extended to allow a different listener to be associated
with an individual transaction, but I chose not to do that yet.