Fixes #1431: SizeStatisticsCollectorTest.tryStaleRead sometimes fails with "version not valid"
The "version not valid" error message appears to happen in `tryStaleRead` if the stale read version given is negative (or zero). That should essentially never happen on any "real" cluster, but on test clusters that are brought up for running these unit tests, that's a real possibility. This solution just makes sure the commit version of FDB is big enough that the new read version won't cause an error, and it skips the test if that assumption would be broken, as failing the build is worse than skipping this test. It does open us up to the possibility that this test gets skipped entirely (if the cluster is always too new), but that seems relatiely unlikely. It would perhaps be better if we could simulate the cluster being old by injecting a failure at read time, but that's a bit difficult to do with the current set of abstractions.

This fixes #1431.