Fixes #1371: Uniqueness violations during check version appear to be silently ignored (#1373)
* Fixes #1371: Uniqueness violations during check version appear to be silently ignored

This validates now that when a unique index is added with duplicate entries that the build does not succeed, but also that the store opening operation isn't failed. Instead, it just means that the index is placed in an unbuilt state that will need to be resolved later (like all other uniqueness violations). This behavior means that (1) new indexes won't be built which have violations of their own invariants and (2) adding a new unique index does not put the user into a state where they might accidentally make a store impossible to open.

This adds a new package-private class, `IndexUniquenessCommitCheck`, which the `FDBRecordStore` will create when asked to add a new check. It associates the check with the given index so that the pertinent checks for each index can be collected and waited on when needed.