Begin #1422: Rename variables called "record"
During some down time, I began removing some of the places we call a variable "record" in anticipation of Java 17 making that a reserved word. I suspect we'll end up wanting to break this one up over many PRs given the number of instances of this that we need. This handles the `FDBRecord` interface and its implementations. In those classes, I turned `record` into `protoRecord` if it was a `Message` type (or subclass), which also distinguishes it from the wrapper "record". In a few places, it became `rec` if the `record` in question was already a wrapper.