Clean up LuceneDocumentFromRecord
The main addition is a new class, LuceneIndexExpressions, which implements in a single place knowledge of what components are allowed in a Lucene index key expression and how they nest to inform document creation. This same class walks the tree either against FDBRecords to make Lucene document fields or against Protobuf meta-data to symbolically represent what that _would_ do.

The old code had an implementation for the next two, but they were never closed:
* Resolves #1235: Lucene should support more field data types
* Resolves #1297: Lucene indexing of map-like fields
This new scheme bases the mapping on explicit function calls, rather than heuristics.
The default index on a concatenation of fields will have single tokens, not full text.

This only minimally addresses the Lucene planner: just enough to keep the tests running.

This eliminates the need for special Lucene key expressions (other than the functions), which in turn eliminates the need for special serialization post-processing. That entire mechanism is removed (a core change).