Resolves #1493: Consider replacing single record range deletes with multiple single key deletes
This changes the single-record delete path so that in cases where we know which key(s) the record used (including taking split points into account), we issue a series of single-key deletes rather than one clear range over the whole record. As issue #1493 descrbies in more detail, this approach should be more efficient for the currently-in-development RocksDB storage engine (as well as any other LSM storage engine that might one day be used) while being more-or-less a wash for the other storage engines.

This resolves #1493.