Resolves #1494: Update log4j version to avoid RCE vulnerability
This updates the version of log4j used by our tests and example code to 2.16.0, which removes the JNDI logic and thus patches the RCE vulnerability. As noted in the issue, the main library only exposes slf4j, so an adopter is only vulnerable if they are using the log4j logging implementation with a vulnerable version, but our own tests and example logic was technically vulnerable, though neither exposed remote endpoints.

This resolves #1494.