Reduce cognitive complexity of ComposedBitmapIndexAggregate.indexScan
In response to:
https://sonarcloud.io/project/issues?id=FoundationDB_fdb-record-layer&open=AX7G6zY5-M4bmnrO40tL&pullRequest=1523&resolved=false&types=CODE_SMELL
I pulled out a couple methods. The bit that I wanted to pull out
was the entire code to "splice" the key, but I couldn't really
figure out what it was trying to do, semantically, so I extracted
a couple, more straightforward methods that I could name.
This rule is probably the most important when adding new code, and
the most dangerous when asking people to change adjacent code.