Resolves #1520: Expose IndexQueryabilityFilter for Aggregate planning (#1523)
* Resolves #1520: Expose IndexQueryabilityFilter for Aggregate planning

For IndexFunctionHelper and ComposedBitmapIndexAggregate, this change
adds it as a required parameter.

Since FDBRecordStoreBase is MAINTAINED this adds overloads for:
- getSnapshotRecordCountForRecordType
- evaluateAggregateFunction
that take the IndexQueryabilityFilter, using TRUE to maintain
backwards compatibility. The old method is left as a default
method, requiring implementers to add the new method.

* Address easy comments from PR #1523

* As suggested by sonarqube, remove public from tests

See: https://sonarcloud.io/organizations/foundationdb/rules?open=java%3AS5786&rule_key=java%3AS5786

* Reduce cognitive complexity of ComposedBitmapIndexAggregate.indexScan

In response to:
https://sonarcloud.io/project/issues?id=FoundationDB_fdb-record-layer&open=AX7G6zY5-M4bmnrO40tL&pullRequest=1523&resolved=false&types=CODE_SMELL
I pulled out a couple methods. The bit that I wanted to pull out
was the entire code to "splice" the key, but I couldn't really
figure out what it was trying to do, semantically, so I extracted
a couple, more straightforward methods that I could name.
This rule is probably the most important when adding new code, and
the most dangerous when asking people to change adjacent code.

* Add IndexQueryabilityFilter to other aggregate methods in FDBRecordStoreBase

* Cleanup FDBRestrictedIndexQueryTest a bit