Rough pass at extracting LimittedRunner
In essence, this is an attempt to abstract away IndexingThrottle
so that it could be used for other cross-transaction work that
may want to limit the amount of work that it is touching. One
such example of that would be to Export/Import mechanism, which
would want to do as much work per-transaction as it can, but would
need to save across many transactions.