Fix IndexingByRecords.buildRanges rangeDeque management
At first it was incorrectly ending if it failed to commit, as it
wouldn't add back the range that didn't commit.
But we also need to make sure that RecordBuiltRangeException
results in the queue being correctly filled, and not adding back
an incorrect range.
This code is really complicated, and I'm a little concerned, but
supposedly we're removing it soon.