Fixes #1563: Record-type prefix check failures can cause deleteRecordsWhere failures
The primary change here is an adjustment to the way that the `Key.Expressions.hasRecordTypePrefix` function works so that it can function if the key expression would otherwise require a split at a column that is not expressible with the key expression language. In particulr, it now uses the key-with-value expression's whole key instead of just the "key expression" when making the prefix determination, with some care taken so as to not accidentally allow a value-only key expression to pass.

Then, it turns out there were some other potential failures with deleteRecordsWhere and such a key with value expression, this time during matching evaluation. In particular, the key is again split apart to try and determine if the key expression satisfies the delete-records-where filter. However, if the split happens at a bad place, it would just give up and say "no match here". However, this can be made less conservative by continuing to try and match a prefix of the key expression with the query, and if that prefix matches, then we are also good.

This fixes #1563.