Resolves #1575: Lucene FDBDirectory could cache listAll results
This restructures the way the `FDBDirectory` handles caching the file name to `FDBLuceneFileReference` map. Previously, there was a cache where entries were probably filled by a call to `listAll()` that would happen when the directory was opened anyway, but now, this keeps a full cache of file names and file reference information in memory for the directory. In theory, this could get too big if there are too many files, but the calling code relies on being able to list all files anyway, so at least for now, that's a problem either way. Note that the calling code sometimes makes multiple `listAll()` calls in the same request, and so this cache can speed up calls after the first one.

This also tweaks the way that the list is read from the database in that it uses the `WANT_ALL` streaming mode instead of the (default) `ITERATOR` streaming mode. This streaming mode tries to optimize the number of round trips required to complete a range read, which means that the time to first byte can be longer but the time to the last byte can be shorter. Given that the full list is needed in this case before we can do anything useful with it, optimizing for time to last byte should be an improvement in this case.