Fixes #1583: Indexes with groups erroneously accept deleteRecordsWhere queries on grouped columns
This adjusts the methods used by `canDeleteWhere` on indexes defined on a `GroupingKeyExpression` so that it tries to satisfy the query from the grouping sub key rather than the whole key to limit the number of accepted `deleteRecordsWhere` predicates to more closely align with that is actually possible to safely delete. This includes some regression tests that hit the error, as well as a small amount of refactoring that is to set up #1580.

This fixes #1583.