Resolves #1580: Lucene index maintainer should support delete where on group key
This adds support for `deleteRecordsWhere` to the `LuceneIndexMaintainer`. To do so, it needed to refactor a fair amount of the infrastructure around session management of `FDBDirectory`. This is because the `FDBDirectory` maintains a cache of the data inside of it, which means that if there's a range delete and then a read in the same transaction, the in-memory cache could result in deleted data being returned. The new code puts one object per index, the `FDBDirectoryManager`, in the transaction-level session cache, and then the `LuceneIndexMaintainer` uses it to grab a directory, or (in the case of a range delete) to invalidate directories in the cache. This also coalesced various commit checks to close these resources into a single commit check on the directory manager, which closes all oustanding objects.

This resolves #1580.

As part of the testing, I needed to be able to read data on an empty index, which through an error (see #1597). This adds logic to translate the error into an empty cursor, so this fixes #1597.