Fixes #1601: Make EvaluationContextBuilder API backwards compatible
This makes the `TypeRepository` handling in the `EvaluationContext` API more-buildery, with a default value of `TypeRepository.EMPTY_SCHEMA`. This also simplifies some logic that was creating a child builder off of a parent context and then had to remember to pass in the parent's `TypeRepository` to the new child, as that's now handled by the `Builder` itself, though it is still possible for the child to override the `TypeRepository` if it wishes. This is a breaking API change of of 3.1.248.0, though it reverts it back to something that is backwards compatible with 3.1.247.0.

This fixes #1601.