Resolves #1577: Lucene FDBDirectory should cache file counter
This adds a new `AtomicLong` to the `FDBDirectory` where the file sequence counter is cached. The first time a user tries to get an increment, the value is loaded from the database and put into the cache. All subsequent accesses can use that cache, and the `AtomicLong` guarantees thread safety. Note that the FDB read-your-writes cache was already keeping a cached copy, but this removes the extra JNI hop and requirement to go through the FDB network thread in order to get a new increment.